Fixed a bug in the demo.py file
The bug was that he image keypoints were being drawn on the wrong point on the image when stream lit draw method was used on them.
This was fixed by using pyplot method inorder to draw out on the image.
The keypoints were drawn out in tuples of 2, which was creating some kind of conflict with the ouput it was fixed by using kp_x and kp_y instead.
The facial_kp_detection class was also modified to include the new changes which allowed it to work with only cpu device.